CustomActions: Screen off availability
Ensure some actions work while the screen is off (Assistant, Camera,
Flashlight, Clear notifications) and the actions that don't, can't be
called on.

Voice search can't be launched from within the keyguard, requires the
screen to be turned on. With that said, doesn't make sense to have it
triggered while the screen is turned off. Volume panel, Notifications,
Screenshot and QS panel are the same way.

Screenshot 'technically' works while the screen is off but taking a
screenshot while the screen is off gives you a black and useless screenshot.
Nobody wants or needs that.

'Screen off' action for obvious reasons doesn't work while the screen is
off. If you need me to explain this one to you, WOW....

Change-Id: I99af6eb29a9fabdd3ff617dc018c0d0204b8aeb6
Signed-off-by: spezi77 <spezi7713@gmx.net>